feat: 🎸 Add logic in webstore for refreshing auth tokens

+++ apps/webstore/src/app/shared/services/authentication/auth.service.ts
@@ -18,6 +18,8 @@ const httpOptions = {
 export class AuthService {
   constructor(private http: HttpClient, private localStorageService: LocalStorageService, private router: Router) {}
 
+  private refreshingTokens: boolean = false;
+
   /**
    * Perform a login request to the API
    * @param params user credentials
@@ -52,6 +54,7 @@ export class AuthService {
     );
   }
 
+  // TODO - send a logout request
   /**
    * Remove the user authentication information from local storage
    */
@@ -60,6 +63,30 @@ export class AuthService {
     // the tree design should no longer be accessible
     this.localStorageService.removeItem(LocalStorageVars.designFamilyTree);
     this.router.navigate(['/home']);
+  }
+
+  /**
+   * Performs a query to the backend to refresh the access token.
+   * Updates the data in local storage.
+   */
+  refreshAccessToken(): void {
+    // check if the tokens are already being refreshed
+    if (this.refreshingTokens) {
+      return;
+    }
+    this.refreshingTokens = true;
+    // call the api to refresh the authentication info
+    this.http.get<IAuthUser>(`${env.apiUrl}/auth/refresh`).subscribe({
+      next: (authUser: IAuthUser) => {
+        this.localStorageService.setItem(LocalStorageVars.authUser, authUser);
+        this.refreshingTokens = false;
+        return authUser;
+      },
+      error: (error) => {
+        console.error(error);
+        this.refreshingTokens = false;
+      },
+    });
   }
 
   /**
@@ -95,8 +122,16 @@ export class AuthService {
         return false;
       }
       const isExpired = this.isJwtExpired(accessToken);
+      // if the access token is expired, check the refresh token
       if (isExpired) {
-        console.warn('Your session has expired, logging you out');
+        const refreshToken = authUser.getValue()?.refreshToken;
+        if (refreshToken === undefined) {
+          return false;
+        }
+        if (!this.isJwtExpired(refreshToken)) {
+          return true;
+        }
+        console.warn('Valid: Your session has expired, logging you out');
         this.localStorageService.removeItem(LocalStorageVars.authUser);
       }
       return !isExpired;
